Organize Test folder (#50)
