+++ Tests/Business/Mokups/mokUserAlice.cs
@@ -3,7 +3,7 @@ using eCommerce.Business;
 using eCommerce.Business.Service;
 using eCommerce.Common;
 
-namespace Tests.Mokups
+namespace Tests.Business.Mokups
 {
     public class mokUserAlice : IUser
     {
