+++ Tests/Business/UserManagerTests/UserAuthMock.cs
@@ -1,11 +1,9 @@
-﻿using System.Collections;
-using System.Collections.Concurrent;
+﻿using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Runtime.InteropServices;
 using eCommerce.Auth;
 using eCommerce.Common;
 
-namespace Tests.Business
+namespace Tests.Business.UserManagerTests
 {
     /// <summary>
     /// <para>Implementation of concurrent.</para>
