+++ Tests/Business/UserManagerTests/UserManagerTests.cs
@@ -1,12 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using eCommerce.Auth;
 using eCommerce.Business;
 using eCommerce.Common;
 using NUnit.Framework;
 
-namespace Tests.Business
+namespace Tests.Business.UserManagerTests
 {
     [TestFixture]
     public class UserManagerTests
